backend/pyproject.toml
[project]
name = "manga-translator-backend"
version = "0.1.0"
description = "Manga translation pipeline backend"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Settings & validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.5.0",

    # Image processing
    "Pillow>=10.4.0",
    "opencv-python-headless>=4.10.0",
    "numpy>=1.26.0",

    # Text detection
    "craft-text-detector>=0.4.3",

    # OCR
    "paddleocr>=2.8.0",
    "paddlepaddle>=2.6.0",

    # Translation
    "openai>=1.50.0",

    # Inpainting
    "simple-lama-inpainting>=0.1.0",

    # Archive handling
    "rarfile>=4.2",

    # Utilities
    "httpx>=0.27.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
